Add files via upload
If the file is having problems in downloading then use this link https://colab.research.google.com/drive/1HFNcaz7YkmkqBsCUZfb_fyR1ijenzAFU?usp=sharing and go to the File section in the toolbar (on the top left part of the screen) and download the file in .ipynb format. Then upload this file to your own Google Colab's account and start solving the assignment.